@@ -1,27 +1,20 @@
 package com.example.indydesk;
 
-import android.app.ProgressDialog;
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.TextUtils;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ProgressBar;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
-import com.google.firebase.FirebaseApp;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-
-import butterknife.BindView;
-import butterknife.ButterKnife;
 
 public class SignupActivity extends AppCompatActivity {
 
@@ -35,7 +28,7 @@ public class SignupActivity extends AppCompatActivity {
         private ProgressBar progressBar;
         private FirebaseAuth auth;
         String name,email,address,password,phone;
-        android.support.v7.widget.Toolbar  toolbar;
+        androidx.appcompat.widget.Toolbar  toolbar;
         @Override
         protected void onCreate(Bundle savedInstanceState) {
             super.onCreate(savedInstanceState);
